# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Forge - Website Builder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  draft: "draft",
  published: "published",
  archived: "archived",
} as const;

export interface Project {
  id: number;
  name: string;
  description?: string;
  status: ProjectStatus;
  template: string;
  pageCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectBody {
  name: string;
  description?: string;
  template: string;
}

export type UpdateProjectBodyStatus =
  (typeof UpdateProjectBodyStatus)[keyof typeof UpdateProjectBodyStatus];

export const UpdateProjectBodyStatus = {
  draft: "draft",
  published: "published",
  archived: "archived",
} as const;

export interface UpdateProjectBody {
  name?: string;
  description?: string;
  status?: UpdateProjectBodyStatus;
}

export interface Page {
  id: number;
  projectId: number;
  title: string;
  slug: string;
  content: string;
  order: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePageBody {
  title: string;
  slug: string;
  content?: string;
  order?: number;
}

export interface UpdatePageBody {
  title?: string;
  slug?: string;
  content?: string;
  order?: number;
}

export interface DashboardSummary {
  totalProjects: number;
  publishedProjects: number;
  draftProjects: number;
  totalPages: number;
}