# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Forge - Website Builder API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all projects
 */
export const ListProjectsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  status: zod.enum(["draft", "published", "archived"]),
  template: zod.string(),
  pageCount: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListProjectsResponse = zod.array(ListProjectsResponseItem);

/**
 * @summary Create a new project
 */
export const CreateProjectBody = zod.object({
  name: zod.string(),
  description: zod.string().optional(),
  template: zod.string(),
});

/**
 * @summary Get a project by ID
 */
export const GetProjectParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProjectResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  status: zod.enum(["draft", "published", "archived"]),
  template: zod.string(),
  pageCount: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a project
 */
export const UpdateProjectParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProjectBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().optional(),
  status: zod.enum(["draft", "published", "archived"]).optional(),
});

export const UpdateProjectResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  status: zod.enum(["draft", "published", "archived"]),
  template: zod.string(),
  pageCount: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a project
 */
export const DeleteProjectParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List pages for a project
 */
export const ListPagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListPagesResponseItem = zod.object({
  id: zod.number(),
  projectId: zod.number(),
  title: zod.string(),
  slug: zod.string(),
  content: zod.string(),
  order: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListPagesResponse = zod.array(ListPagesResponseItem);

/**
 * @summary Create a page for a project
 */
export const CreatePageParams = zod.object({
  id: zod.coerce.number(),
});

export const CreatePageBody = zod.object({
  title: zod.string(),
  slug: zod.string(),
  content: zod.string().optional(),
  order: zod.number().optional(),
});

/**
 * @summary Update a page
 */
export const UpdatePageParams = zod.object({
  id: zod.coerce.number(),
  pageId: zod.coerce.number(),
});

export const UpdatePageBody = zod.object({
  title: zod.string().optional(),
  slug: zod.string().optional(),
  content: zod.string().optional(),
  order: zod.number().optional(),
});

export const UpdatePageResponse = zod.object({
  id: zod.number(),
  projectId: zod.number(),
  title: zod.string(),
  slug: zod.string(),
  content: zod.string(),
  order: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a page
 */
export const DeletePageParams = zod.object({
  id: zod.coerce.number(),
  pageId: zod.coerce.number(),
});

/**
 * @summary Get dashboard summary stats
 */
export const GetDashboardSummaryResponse = zod.object({
  totalProjects: zod.number(),
  publishedProjects: zod.number(),
  draftProjects: zod.number(),
  totalPages: zod.number(),
});

/**
 * @summary Get recently updated projects
 */
export const GetRecentProjectsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  status: zod.enum(["draft", "published", "archived"]),
  template: zod.string(),
  pageCount: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetRecentProjectsResponse = zod.array(
  GetRecentProjectsResponseItem,
);