# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Forge - Website Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalProjects: number;
  publishedProjects: number;
  draftProjects: number;
  totalPages: number;
}