# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Forge - Website Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectStatus } from "./projectStatus";

export interface Project {
  id: number;
  name: string;
  description?: string;
  status: ProjectStatus;
  template: string;
  pageCount: number;
  createdAt: Date;
  updatedAt: Date;
}