# lib/api-zod/src/generated/types/updatePageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Forge - Website Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePageBody {
  title?: string;
  slug?: string;
  content?: string;
  order?: number;
}